[AIP-21] Move operators.s3_file_transform_operator airflow.operators.s3_file_transform_operator
